Add dynamic heap allocator for VibeRTOS with multiple backends
- Introduced a new header file `heap.h` defining the API for the heap allocator.
- Implemented three allocation strategies: First-Fit, Best-Fit, and TLSF (Two-Level Segregated Fit).
- Created a CMake configuration for the heap library, allowing selection of the allocator via Kconfig.
- Developed the First-Fit allocator in `heap_first_fit.c`, managing memory with a simple linked list.
- Developed the Best-Fit allocator in `heap_best_fit.c`, optimizing for fragmentation at the cost of search time.
- Implemented the TLSF allocator in `heap_tlsf.c`, providing O(1) allocation and deallocation.
- Added a private header `heap_priv.h` for shared definitions between allocators.
- Included a stub implementation `heap_stub.c` for cases where heap memory is disabled.
- Enhanced diagnostics with `vibe_heap_stats()` to report memory usage and allocation statistics.

+++ samples/blinky/main.c
@@ -34,7 +34,7 @@ static void blinky_thread_entry(void *arg)
     /* Configure LED pin as output, no pull. */
     vibe_err_t err = gpio_configure(gpio_dev, LED_PIN, GPIO_DIR_OUTPUT, GPIO_PULL_NONE);
     if (err != VIBE_OK) {
-        vibe_printk("[blinky] gpio_configure failed: %d\n", err);
+        vibe_printk("[blinky] gpio_configure failed: %ld\n", (long)err);
         return;
     }
